Make the read-metadata.pl example more robust
Sporadically, we will encounter an "error" message (about a missing sink
element) before a "tag" message even for a valid file.  So instead of
aborting on an "error" message, just continue.  Also, shorten the wait
time for state changes from 5 to 0.5 seconds, and improve the error
message for unknown file types.